# InnovatEPAM-Portal/specs/main/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: InnovatEPAM Portal API
  version: 0.1.0
  description: Core authentication, idea submission, listing, draft, and review endpoints for the InnovatEPAM Portal feature.
servers:
  - url: http://localhost:3000
paths:
...
  /api/ideas:
    get:
      summary: List ideas for the current user or privileged reviewer
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Idea list returned
        '401':
          description: Missing or invalid token
    post:
      summary: Submit a new idea with optional attachment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, category]
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Idea submitted
        '400':
          description: Validation failure
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT